requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-rest-auth
channels
requests
gunicorn